requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.0.0
fastjsonschema==2.16.1
ffmpeg-python==0.2.0
filelock==3.11.0
Flask==3.0.0
flatbuffers==22.10.26
fpdf==1.7.2
frozenlist==1.3.3
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gcsfs==2023.10.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.23.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.36.4
google-cloud-appengine-logging==1.3.2
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-logging==3.8.0
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
grip==4.6.2
grpc-google-iam-v1==0.12.7
grpcio==1.51.3
grpcio-status==1.59.3
h5py==3.7.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jax==0.4.19
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.3.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
numba==0.56.4
numpy==1.23.4
oauthlib==3.2.2
openai==0.27.4
openai-whisper==20230314
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
path-and-address==2.0.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.2
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-version==0.0.2
pytube==12.1.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.7
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.0
shapely==2.0.2
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.5.0
streamlit==1.21.0
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.15.0
threadpoolctl==3.2.0
tiktoken==0.3.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
virtualenv==20.21.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.2
zipp==3.15.0
